Implemented the cool looking hover ui
The only thing left in this the actual social links of the companies and their description.

The file where the company details for VC and Portfolio is present in /pages/Ventures/companyDetails.js

Please add the social links and if a particular social media site is not used by that company then leave a empty string.

Also add a small description and change the default one.